blog/requirements.txt
alembic==1.4.2
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Sitemap==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
qiniu==7.2.6
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.0
WTForms==2.2.1
zipp==3.1.0
